manga-ocr-server/requirements.txt
fastapi
uvicorn
manga-ocr
pillow
python-multipart
